Generate bindings once for all platforms and architectures
Uses CI based on https://github.com/Smithay/drm-rs/pull/176.

Includes https://github.com/Smithay/gbm.rs/pull/29.
